requirements.txt
pandas>=2.0.0
scikit-learn>=1.3.0
datasets>=2.14.0
nltk>=3.8.0
google-genai>=0.8.0
anthropic[vertex]>=0.39.0
requests>=2.31.0
google-auth>=2.30.0
python-dotenv>=1.0.0
openpyxl>=3.1.0